fixed pathfinder

+++ EUS/AStarPathfinder.h
@@ -27,7 +27,7 @@ private:
 
 	MapGrid& grid;
 
-	bool isWalkable(const size_t indexX, const size_t indexY, const std::vector<NodeInfo>& nodeInfos);
+	bool isWalkable(const int indexX, const int indexY, const std::vector<NodeInfo>& nodeInfos);
 	void internalFindPath(bool& foundPath, std::vector<pmath::Vec2f>& outPath, const std::vector<NodeInfo>& nodeInfos, const pmath::Vec2f& start, const pmath::Vec2f& goal);
 	void extractVectorFrom(const MapNode& current, std::vector<pmath::Vec2f>& outPath);
 public:
